Implement "low-level" key press event
This is separate to the existing key press event, mainly because they
are also different things in scratch. For example, there are different
blocks that listen to each.

One "missing" aspect in the implementation is making it synchronous.
However, there is no concept of threads that are waiting on this event
explicitly (but it can happen, e.g. a loop with a key press sensing
block as condition). It is almost impossible to find these reliable
without existing those threads. For example, just waiting on all threads
with the key sensing block is not sufficiënt: the thread may have
passed that block and can be in another loop.

Additionally, at the moment I think in most constructs that use this
event, waiting on all threads will not be that useful. #15 will be much
more useful.