Listen to "stop all" blocks to detect terminated threads
Some context: for various synchronous operations, we maintain a list of
threads. We then check this list on each thread update, and determine
which of those threads ended. This allows us to resolve the thread
listener.

However, when using the "stop all scripts" block, or a
"stop all scripts for x" block, the threads are not stopped during the
step function.

We solve this by also listening to the "stopThread" method.